Spotbugs(medium) : NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE

+++ jpsonic-main/src/main/java/org/airsonic/player/service/LastFmCache.java
@@ -130,10 +130,13 @@ public class LastFmCache extends Cache {
 
     @Override
     public void clear() {
-        for (File file : cacheDir.listFiles()) {
-            if (file.isFile()) {
-                if (!file.delete() && LOG.isWarnEnabled()) {
-                    LOG.warn("The file '{}' could not be deleted.", file.getAbsolutePath());
+        File[] listFiles = cacheDir.listFiles();
+        if (listFiles != null) {
+            for (File file : listFiles) {
+                if (file.isFile()) {
+                    if (!file.delete() && LOG.isWarnEnabled()) {
+                        LOG.warn("The file '{}' could not be deleted.", file.getAbsolutePath());
+                    }
                 }
             }
         }
